requirements.txt
# This file was autogenerated by uv via the following command:
#    uv pip compile pyproject.toml
argh==0.31.3
    # via eta-kappa-notes (pyproject.toml)
attrs==25.4.0
    # via
    #   jsonschema
    #   referencing
contourpy==1.3.3
    # via matplotlib
cycler==0.12.1
    # via matplotlib
fonttools==4.60.1
    # via matplotlib
jsonschema==4.25.1
    # via eta-kappa-notes (pyproject.toml)
jsonschema-specifications==2025.9.1
    # via jsonschema
kiwisolver==1.4.9
    # via matplotlib
matplotlib==3.10.7
    # via
    #   eta-kappa-notes (pyproject.toml)
    #   mpl-tools
mpl-tools==0.4.1
    # via eta-kappa-notes (pyproject.toml)
numpy==2.3.4
    # via
    #   eta-kappa-notes (pyproject.toml)
    #   contourpy
    #   matplotlib
packaging==25.0
    # via matplotlib
pillow==12.0.0
    # via matplotlib
pyparsing==3.2.5
    # via matplotlib
python-dateutil==2.9.0.post0
    # via matplotlib
pyyaml==6.0.3
    # via eta-kappa-notes (pyproject.toml)
referencing==0.37.0
    # via
    #   jsonschema
    #   jsonschema-specifications
rpds-py==0.28.0
    # via
    #   jsonschema
    #   referencing
six==1.17.0
    # via python-dateutil
typing-extensions==4.15.0
    # via referencing
